requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
cffi==2.0.0
exceptiongroup==1.3.1
fastapi==0.124.4
filelock==3.19.1
flatbuffers==25.12.19
fsspec==2025.10.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
pillow==11.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
sounddevice==0.5.3
starlette==0.49.3
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
mediapipe==0.10.5